pyproject.toml
[tool.poetry]
name = "clicking"
version = "0.1.0"
description = ""
authors = ["Sarath S Menon <sarathmenon.downey@gmail.com>"]
readme = "README.md"
# package-mode = false

packages = [
    { include = "clicking/api", from = "bases" },
    { include = "clicking/gradio_app", from = "bases" },
    { include = "clicking/vision_model", from = "components" },
    { include = "clicking/output_corrector", from = "components" },
    { include = "clicking/pipeline", from = "components" },
    { include = "clicking/prompt_manager", from = "components" },
    { include = "clicking/prompt_refinement", from = "components" },
    { include = "clicking/dataset_creator", from = "components" },
]

[tool.poetry.dependencies]
python = "^3.10.12"
torch = "^2.4.0"
torchvision = "^0.19.0"
matplotlib = "^3.9.2"
transformers = "^4.44.0"
timm = "^1.0.8"
einops = "^0.8.0"
google-cloud-storage = "^2.18.2"
scipy = "^1.14.0"
pycocotools = "^2.0.8"
fastapi = "^0.112.0"
uvicorn = "^0.30.6"
gradio = "^4.41.0"
openapi-python-client = "^0.21.2"
opencv-python = "^4.10.0.84"
gradio-log = "^0.0.4"
python-dotenv = "^1.0.1"
litellm = "^1.43.18"
promptdown = "^0.7.5"
accelerate = "^0.33.0"
huggingface-hub = "^0.24.6"
setuptools = "^73.0.1"
wheel = "^0.44.0"
wandb = "^0.17.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ipykernel = "^6.29.5"

# [tool.poetry.scripts]
# init = "sh init_script.sh"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
